Allow any object input
All the input to be any object. After validating that the string is not undefined or null, it is then sent to the toString method to convert it to a string. After it is a string it is then validated.
This allows the validation function to be used in more than just string only situations.