Refactor Authentication Library
Update the library to be able to key vault authenticate to grab a secret and build a token chain intelligently. It will be able to auto detect if it needs to grab a secret from a keyvault or if it will use the environmental variables.

Script now returns async object instead of a sync object. added method for kv to process some data seperately from the constructor.