Fix auth bug
Fix a bug in authentication that the chained credential is unable to build a managed identity based authentication because the client secret could be undefined, and that the client id and tenant id could be set to "none" instead of an empty string. An empty string is best and won't kill the auth process.