Update the SC Patch
Update the settings catalog patch debug route so that it takes advantage of the new put style method that the graph client now uses.
Remove the /settingscatalog/id/settings route as it is no longer needed since the patch command uses the put method in the client and doesn't have the traversal issues.