Refactor Get/Update MEM Tag Mmethods
Renamed the Endpoint Scope Tag methods to be MEMScopeTag instead, this follows the acronym usage throughout the graph client and reduces the excessive namespace length.
Refactored the get and update methods of the MEM Tag methods to be best practice and add additional functionality.
Ran document format against the graph client and debug routes too.