Initial Core Router
Built the initial core router file.
this includes a configuration data structure initialization and a template for the Assign PAW route.
This file is meant to be the business logic REST API for the UI to connect to.
It will not expose the low level methods that the debug router exposes.