Add Unconfiged Tag Support
Add the ability for the config engine to identify an unconfiged/blank role scope tag description in the config parser. This usually is the case for new deployments where the deployment admin makes a role scope tag ahead of time instead of using the deployment engine to deploy the role scope tag.