Fix 3 bugs + Naming
Fix the scope tag assignment bug by changing it to use the scope tag ID instead of the name.
Fix the CommissionedDate to use the proper format instead of the raw toString that it was automatically using.
Corrected the device group membership to notify the graph client it is giving it a device id and not an object ID.
Updated the device group naming to prefix the device group with "PAW - " to have a better naming standard to easily see at a glance which groups are PAW groups.