Change to JSON
Change the OpenAPI spec to a raw JSON file instead of a TS file so that it conforms to the OpenAPI 3.x spec. The JSON file is static and the loader now dynamically injects data in at runtime instead of dynamically making an object then feeding that dynamic object into the runtime.