Add firewall autoconfig & fix bug
Add a firewall auto config step to allow the app to be accessible.
Fixed a bug with the user installer preventing the app to install properly. This was caused by using an encrypted property. Switched to normal property.
Rename the Scheduled task to be easier to find in the task scheduler mgmt console.