Update Device Models
Rename the models to be more generic as the scope of the project has expanded since the inception of the project's original app code.
PAW device is now PSM Device (since dev workstations will be covered)
Device Item is now Psm Autopilot Device as the device item is too generic for its use case.
Since the full parsed data is now available, the property names of the app have also been updated to reflect the proper data structures.