Refactor Auth Engine
Refactored the auth engine to simplify the authentication engine and fix a massive architectural issue that prevented all but local var auth from working.
Simplified data validators by splitting the config validation out of the data validation section. Now the data imput validators are more simple to understand and the config validation is completed elsewhere as it is not as critical to check and can be done at execution time instead of before.